# spatialguard/packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SpatialGuard",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/test-videos/{video_id}/media": {
      "get": {
        "summary": "Test Video Media",
        "operationId": "test_video_media_v1_test_videos__video_id__media_get",
        "parameters": [
          {
            "name": "video_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Video Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/test-videos/{video_id}/track": {
      "get": {
        "summary": "Test Video Track",
        "operationId": "test_video_track_v1_test_videos__video_id__track_get",
        "parameters": [
          {
            "name": "video_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Video Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestVideoTrack"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/test-videos/{video_id}/classify": {
      "post": {
        "summary": "Classify Test Video",
        "operationId": "classify_test_video_v1_test_videos__video_id__classify_post",
        "parameters": [
          {
            "name": "video_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Video Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentClassification"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IncidentClassification": {
        "properties": {
          "label": {
            "type": "string",
            "enum": [
              "delivery_activity",
              "face_covering_visible",
              "possible_weapon_visible",
              "possible_unauthorized_entry",
              "unidentified_person",
              "package_visible",
              "animal_visible",
              "vehicle_visible",
              "unclear",
              "no_relevant_activity"
            ],
            "title": "Label"
          },
          "display_label": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Display Label"
          },
          "confidence": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high"
            ],
            "title": "Confidence"
          },
          "summary": {
            "type": "string",
            "maxLength": 300,
            "minLength": 1,
            "title": "Summary"
          },
          "visible_evidence": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 5,
            "title": "Visible Evidence"
          },
          "uncertainty": {
            "type": "string",
            "maxLength": 300,
            "minLength": 1,
            "title": "Uncertainty"
          },
          "model": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Model"
          },
          "response_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Response Id"
          },
          "analyzed_at": {
            "type": "string",
            "title": "Analyzed At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "label",
          "display_label",
          "confidence",
          "summary",
          "visible_evidence",
          "uncertainty",
          "model",
          "analyzed_at"
        ],
        "title": "IncidentClassification"
      },
...
      "TestTrackPoint": {
        "properties": {
          "t_seconds": {
            "type": "number",
            "minimum": 0.0,
            "title": "T Seconds"
          },
          "foot_x_norm": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Foot X Norm"
          },
          "foot_y_norm": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Foot Y Norm"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "t_seconds",
          "foot_x_norm",
          "foot_y_norm",
          "confidence"
        ],
        "title": "TestTrackPoint"
      },
...
      "TestVideoTrack": {
        "properties": {
          "video_id": {
            "type": "string",
            "title": "Video Id"
          },
          "detector": {
            "type": "string",
            "title": "Detector"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/TestTrackPoint"
            },
            "type": "array",
            "title": "Points"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "video_id",
          "detector",
          "points"
        ],
        "title": "TestVideoTrack"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}